2.3.15: fix %zu logging (portable casts) + relieve internal-SRAM OOM in playback
(1) %zu printed literal "zu" — ESP-IDF newlib-nano drops %z, which was hiding
the render/drop/queue counters in serial. Fixed PROPERLY (not the global
CONFIG_NEWLIB_NANO_FORMAT flag): every site → %lu + (unsigned long) cast.
Files: video_player.cc (8), media_player.cc, websocket_protocol.cc,
ogg_demuxer.cc, jpeg_to_image.c.

(2) Internal SRAM dipped to ~255 B during video playback (the real cause of the
intermittent sync_http_open_failed — heap OOM, not sockets). Can't avoid internal
SRAM entirely (LCD SPI DMA + I2S audio DMA need internal DMA-capable RAM), but
the headroom was being eaten by small transient allocs (audio packets ~960B,
vectors, JSON) under CONFIG_SPIRAM_MALLOC_ALWAYSINTERNAL=2048. Lowered to 512 so
they land in PSRAM. sdkconfig edited in place (board type untouched).

+++ main/display/lvgl_display/jpg/jpeg_to_image.c
@@ -58,8 +58,8 @@ static esp_err_t decode_with_new_jpeg(const uint8_t* src, size_t src_len, uint8_
     size_t required_len = out_info.width * out_info.height * 2;
     if (preallocated_out != NULL) {
         if (preallocated_out_len < required_len) {
-            ESP_LOGE(TAG, "Preallocated JPEG output buffer too small: need=%zu got=%zu",
-                     required_len, preallocated_out_len);
+            ESP_LOGE(TAG, "Preallocated JPEG output buffer too small: need=%lu got=%lu",
+                     (unsigned long)required_len, (unsigned long)preallocated_out_len);
             ret = ESP_ERR_INVALID_SIZE;
             goto jpeg_dec_failed;
         }
@@ -187,8 +187,8 @@ static esp_err_t decode_with_hardware_jpeg(const uint8_t* src, size_t src_len, u
 
     if (preallocated_out != NULL) {
         if (preallocated_out_len < out_buf_len) {
-            ESP_LOGE(TAG, "Preallocated hardware JPEG output buffer too small: need=%zu got=%zu",
-                     out_buf_len, preallocated_out_len);
+            ESP_LOGE(TAG, "Preallocated hardware JPEG output buffer too small: need=%lu got=%lu",
+                     (unsigned long)out_buf_len, (unsigned long)preallocated_out_len);
             ret = ESP_ERR_INVALID_SIZE;
             goto jpeg_hw_dec_failed;
         }
@@ -211,7 +211,7 @@ static esp_err_t decode_with_hardware_jpeg(const uint8_t* src, size_t src_len, u
     ESP_LOGD(TAG, "Expected %d bytes, got %" PRIu32 " bytes", out_buf_len, out_size);
 
     if (out_size != out_buf_len) {
-        ESP_LOGE(TAG, "Decoded image size mismatch: Expected %zu bytes, got %" PRIu32 " bytes", out_buf_len, out_size);
+        ESP_LOGE(TAG, "Decoded image size mismatch: Expected %lu bytes, got %" PRIu32 " bytes", (unsigned long)out_buf_len, out_size);
         ret = ESP_ERR_INVALID_SIZE;
         goto jpeg_hw_dec_failed;
     }
@@ -229,7 +229,7 @@ static esp_err_t decode_with_hardware_jpeg(const uint8_t* src, size_t src_len, u
             out_buf[2 * i] = rgb565 & 0xFF;
         } while (i != 0);
         out_size = header_info.width * header_info.height * 2;
-        ESP_LOGD(TAG, "Converted GRAY8 to RGB565, new size: %zu", out_size);
+        ESP_LOGD(TAG, "Converted GRAY8 to RGB565, new size: %lu", (unsigned long)out_size);
     }
 
     ESP_LOG_BUFFER_HEXDUMP(TAG, out_buf, MIN(out_size, 256), ESP_LOG_DEBUG);
